working on itteration 1
I have implemented "read"-method in "user", you only need to send the
name of the patient . Patient record lie in
"files/PatientRecords/".Added a method "isReadRequestAvailable(String
FILENAME)" witch is called at the begining of "read", this is to be
implemented for each user class. it makes sure that the user doesn´t
request reading access to files they arent allowed to access, should
other parts of the system fail.