reading of files
I have worked on the reading of files. we now read nurses, doctors, govs
and divisions. I also moved the "hub" of the server to a new class
"Hub", with the intention to use "Server" for connection only and not
having to fight with "static" in a main class. this can however be moved
to any of the others. Althou I have yet to implement it, the acctual
construktion of users have been moved to said classes., them building
themselves from name only.